==> reports/2_CommonGardenData.qmd <==
---
title: "Climatic data in the common gardens"
date: last-modified
format: 
  html:
    toc: true
    toc-depth: 4
    code-fold: true
    page-layout: full
embed-resources: true
bibliography: references.bib
editor_options: 
  chunk_output_type: console
---

<style type="text/css">
body {
   font-size: 15px;
}
code.r{
  font-size: 11px;
}
pre {
  font-size: 11px
}

table {
  font-size: 11px
}
</style>

```{r setup, include=F}
knitr::opts_chunk$set(dev.args = list(png = list(type = "cairo")))
knitr::opts_chunk$set(cache=F, size="tiny")
options(width = 300)
library(knitr)
library(kableExtra)
library(tidyverse)
library(here)

# my own function for building tables in reports
source(here("scripts/functions/kable_mydf.R"))
```

In this report,  we generate a dataset with the average of the climatic variables at the location of the common gardens between the planting and the measurement date.

# Extraction of the climatic data

Climatic data at the locations of the common gardens was extracted with the [Climate Downscaling Tool (ClimateDT)](https://www.ibbr.cnr.it/climate-dt/).

We first prepare a file in the format required by ClimateDT to extract climate data.

```{r GeneratePreExtractionFile}
coord <- read_csv(here("data/CommonGardenData/CommonGardenCoordinates.csv"), show_col_types = FALSE)
  
# As we do not have elevation data for the CG, we use the ClimateDT elevation data 
coord %>% 
  dplyr::select(site,latitude,longitude) %>% 
  arrange(site) %>% 
  mutate(elevation=case_when(site=="asturias" ~ 449,
                             site=="caceres" ~ 290,
                             site=="bordeaux" ~ 58,
                             site=="madrid" ~ 598,
                             site=="portugal" ~ 865)) %>% 
  write_csv(here("data/ClimaticData/CommonGardens/CommonGardensPreExtraction.csv"))
```

::: {.callout-note}
##### DRYAD repository
The annual climatic data downloaded from ClimateDT at the location of the common gardens between 2010 and 2018 is included in the DRYAD repository: file `ClimateDT_CommonGardens.csv` (shown below).
We leave in this dataset all the climatic variables provided by ClimateDT except `MOP` and `AOP` because their full names, units and calculation methods were not specified on the ClimateDT website.
:::

```{r DryadRepoCommonGardenClimaticData}
df <- read_csv(here("data/ClimaticData/CommonGardens/ClimateDT_outputs.csv"), show_col_types = FALSE) %>% 
   dplyr::rename(cg=ID,
                longitude=Longitude,
                latitude=Latitude,
                elevation=Elevation,
                year=Year,
                SP=MSP) %>% 
  dplyr::filter(year %in% 2010:2018) %>% 
  dplyr::select(-MOP, -AOP) # we remove the variables whose full names and other information are not available on the ClimateDT website
  
df %>%  write_csv(here("data/DryadRepo/ClimateDT_CommonGardens.csv"))

df %>% kable_mydf()
```


# Calculating mean climate for each site

Time period of height and survival measurements in each common garden:

  - **Asturias** (Spain) in **March 2014** when the trees were 37 month-old (trees were planted in **February 2011**).
  
  - **Bordeaux** (France) in **November 2018** when the trees were 85 month-old (trees were planted in **October 2011**).
  
  - **Cáceres** (Spain) in **December 2011** when the trees were 8 month-old (trees were planted in **April 2011**). Note that for this common garden, we calculate the bioclimatic variables for the entire year 2011 (instead of calculating the variables only for months between April and December). Indeed, the calculation of the annual bioclimatic variables will be wrong if we do not account for some months, e.g. the mean annual temperature will be higher than expected because we do not account for some winter months.  
  
  - **Madrid** (Spain) in **December 2011** when the trees were 13 month-old (trees were planted in **November 2010**).
  
  - **Fundão** (Portugal) in **May 2013** when the trees were 27 month-old (trees were planted in **February 2011**)


```{r DefineCGmeasurementPeriods}
years_to_select <- list(asturias = c(2011,2014),
                   bordeaux = c(2011,2018),
                   caceres = 2011,
                   madrid = c(2010,2011),
                   portugal = c(2011,2013))

months_to_rm <- list(asturias = list(first_year=c("01"),
                                     last_year=c("04","05","06","07","08","09","10","11","12")),
                     bordeaux = list(first_year=c("01","02","03","04","05","06","07","08","09"),
                                     last_year=c("12")),
                     caceres = list(first_year=NULL,
                                    last_year=NULL),
                     madrid = list(first_year=c("01","02","03","04","05","06","07","08","09","10"),
                                   last_year=NULL),
                     portugal = list(first_year=c("01"),
                                     last_year=c("06","07","08","09","10","11","12")))
```

```{r SetSelectedClimaticVariables}
# Set of climatic variables
# =========================
clim_var <- readRDS(here("data/ClimaticData/NamesSelectedVariables.rds"))
```

We calculate the average of the climatic variables between the planting and the measurement date. We export the dataset for the following analyses.

```{r CalculateMeanClimateForEachCG, warning=F, message=F}
# Function to calculate the average of the climatic variables at the location of the NFI plots
source(here("scripts/functions/calc_avg_clim_var.R"))

clim <- read_csv(here("data/ClimaticData/CommonGardens/ClimateDT_outputs.csv"), show_col_types = FALSE) %>% 
  dplyr::rename(cg=ID,
                longitude=Longitude,
                latitude=Latitude,
                elevation=Elevation,
                year=Year,
                SP=MSP) %>% 
  group_split(cg) %>% 
  purrr::map(\(x){
    
cg <- unique(x$cg)
years_to_select <- years_to_select[[cg]]  
months_to_rm <-  months_to_rm[[cg]]
    
x %>% calc_avg_clim_var(ref_period = years_to_select, id_spatial_points = "cg")

  }) %>% list_rbind() 


clim %>% saveRDS(here("data/ClimaticData/CommonGardens/ClimateCG.rds"))

clim %>% dplyr::select(cg,any_of(clim_var)) %>% kable_mydf
```

We show below the annual climatic data in each common garden for the years between the planting and the measurement date.

```{r ClimateCG}
read_csv(here("data/ClimaticData/CommonGardens/ClimateDT_outputs.csv"), show_col_types = FALSE) %>% 
  dplyr::rename(cg=ID,
                longitude=Longitude,
                latitude=Latitude,
                elevation=Elevation,
                year=Year,
                SP=MSP) %>% 
  group_split(cg) %>% 
  purrr::map(\(x){
  
    
cg <- unique(x$cg)
ref_period <- years_to_select[[cg]]  
months_to_rm <-  months_to_rm[[cg]]

clim_df <- x %>%  
  dplyr::filter(!year %in% c("2041-2070","baseline")) %>% # we remove future climatic data
  dplyr::mutate(year=as.numeric(year)) %>% # year column as numeric so that we can remove years after a given date
  { if(length(ref_period)==1) dplyr::filter(.,year %in% ref_period[[1]]) else # we keep the years of the period of interest
  dplyr::filter(.,ref_period[[1]]<=year& year<=ref_period[[2]]) } %>% 
  dplyr::select(-year) %>%
      dplyr::rename(id=all_of("cg")) # rename the column with ID of the spatial points to 'cg'

clim_df[1,lapply(months_to_rm[["first_year"]], function(x) paste0(c("tmn","tmx","prc"),x)) %>% unlist()] <- NA
clim_df[nrow(clim_df),lapply(months_to_rm[["last_year"]], function(x) paste0(c("tmn","tmx","prc"),x)) %>% unlist()] <- NA

clim_df

  }) %>% 
  list_rbind() %>% 
  kable_mydf
```

# Session information

```{r SessionInfo}
devtools::session_info()
```
